Deduplicate student instance question router
