--- apps/prairielearn/src/pages/studentInstanceQuestionHomework/studentInstanceQuestionHomework.ejs
@@ -1,53 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <%- include('../partials/head'); %>
-    <%- compiled_script_tag('question.ts') %>
-    <script defer src="<%= node_modules_asset_path('mathjax/es5/startup.js') %>"></script>
-    <script>
-      document.urlPrefix = '<%= urlPrefix %>';
-    </script>
-    <% if (question.type != 'Freeform') { %>
-    <script src="<%= node_modules_asset_path('lodash/lodash.min.js') %>"></script>
-    <script src="<%= asset_path('javascripts/require.js') %>"></script>
-    <script src="<%= asset_path('localscripts/question.js') %>"></script>
-    <script src="<%= asset_path(`localscripts/question${effectiveQuestionType}.js`) %>"></script>
-    <% } %>
-    <%- extraHeadersHtml %>
-  </head>
-  <body>
-    <script>
-      // make the file inputs display the file name
-      $(document).on('change', '.custom-file-input', function () {
-          let fileName = $(this).val().replace(/\\/g, '/').replace(/.*\//, '');
-          $(this).parent('.custom-file').find('.custom-file-label').text(fileName);
-      });
-    </script>
-
-    <%- include('../partials/navbar', {navPage: ''}); %>
-    <main id="content" class="container">
-      <div class="row">
-        <div class="col-md-9 col-sm-12">
-          <%- include('../partials/question', {question_context: 'student_homework'}); %>
-        </div>
-
-        <%# Sidebar %> 
-        <div class="col-md-3 col-sm-4 col-xs-12">
-
-          <%- include('../partials/assessmentScorePanel') %>
-
-          <%- include('../partials/questionScorePanel') %>
-
-          <%- include('../partials/questionNavSideButtonGroup') %>
-
-          <% if (config.attachedFilesDialogEnabled) { %>
-              <%- include('../partials/attachFilePanel') %>
-          <% } %>
-
-          <%- include('../partials/instructorInfoPanel') %>
-
-        </div>
-      </div>
-    </main>
-  </body>
-</html>
