+++ docs/dev-guide.md
@@ -857,7 +857,7 @@ To automatically fix lint and formatting errors, run `make format`.
 
 - The core files involved in question rendering are [lib/question-render.js](https://github.com/PrairieLearn/PrairieLearn/blob/master/apps/prairielearn/src/lib/question-render.js), [lib/question-render.sql](https://github.com/PrairieLearn/PrairieLearn/blob/master/apps/prairielearn/src/lib/question-render.sql), and [pages/partials/question.ejs](https://github.com/PrairieLearn/PrairieLearn/blob/master/apps/prairielearn/src/pages/partials/question.ejs).
 
-- The above files are all called/included by each of the top-level pages that needs to render a question (e.g., `pages/instructorQuestionPreview`, `pages/studentInstanceQuestionExam`, etc). Unfortunately the control-flow is complicated because we need to call `lib/question-render.js` during page data load, store the data it generates, and then later include the `pages/partials/question.ejs` template to actually render this data.
+- The above files are all called/included by each of the top-level pages that needs to render a question (e.g., `pages/instructorQuestionPreview`, `pages/studentInstanceQuestion`, etc). Unfortunately the control-flow is complicated because we need to call `lib/question-render.js` during page data load, store the data it generates, and then later include the `pages/partials/question.ejs` template to actually render this data.
 
 - For example, the exact control-flow for `pages/instructorQuestion` is:
